controls dashboard changes

+++ database/migrations/2021_07_25_164407_create_device_settings.php
@@ -19,7 +19,7 @@ class CreateDeviceSettings extends Migration
             $table->decimal('critic_acid')->nullable();
             $table->decimal('pre_filter')->nullable();
             $table->decimal('post_filter')->nullable();
-            $table->decimal('general_service')->nullable();
+            $table->integer('general_service')->nullable();
             $table->timestamps();
         });
     }
